Moved all tests to integration tests
