+++ src/matrix/matrix_op/eig.rs
@@ -204,45 +204,11 @@ where
     }
 }
 
+//-------------------------------------------------------------------------------------------------
+//{{{ mod: tests
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use approx::assert_relative_eq;
-
-    use super::super::super::common::{SMatrixConstructors, DMatrixConstructors};
-
-    #[test]
-    fn test_eigenvalue_decomposition() {
-        let a = SMatrix::<f64, 3, 3>::from_row_slice(&[
-            1.0, 5.0, 0.0,
-            2.0, 4.0, -1.0,
-            0.0, 2.0, 3.0,
-        ]);
-
-        let eig = a.eig().unwrap();
-        // Known eigenvalues for this matrix
-        let expected_eigenvalues = vec![Complex64::new(-0.8595233886152194, 0.0), 
-                                        Complex64::new(5.433664629783286, 0.0), 
-                                        Complex64::new(3.42585875883193, 0.0)];
-
-        for i in 0..3 {
-            assert_relative_eq!(eig.eigvals[i].re, expected_eigenvalues[i].re, epsilon=1e-10);
-            assert_relative_eq!(eig.eigvals[i].im, expected_eigenvalues[i].im, epsilon=1e-10);
-        }
-
-
-        // Known left eigenvectors for this matrix
-        let expected_left_eigenvecotors = SMatrix::<f64, 3, 3>::from_row_slice(&[
-            -0.7212203345550064,  -0.3850687990747861,  -0.3073880480179293,  
-            0.6705630402249634, -0.8536329572455033,  -0.3728399943222721,  
-            0.1737424476304467,  0.3507603088768719,  0.8755015285934659, 
-        ]);
-
-        for i in 0..3 {
-            for j in 0..3 {
-                assert_relative_eq!(eig.left_eigvecs[(i, j)], expected_left_eigenvecotors[(i, j)], epsilon=1e-10);
-            }
-        }
-        
-    }
+mod tests
+{
+  
 }
+//}}}
