+++ src/matrix/matrix_op/lu.rs
@@ -179,92 +179,5 @@ where
 #[cfg(test)]
 mod tests
 {
-
-    use super::super::super::common::{SMatrixConstructors, DMatrixConstructors};
-    use super::SMatrix;
-
-    use approx::assert_relative_eq;
-
-
-    #[test]
-    fn test_non_diagonal_dominant()
-    {
-        let a = SMatrix::<f64, 3, 3>::from_row_slice(&[
-            1.0, 2000.0, 3000.0, 5000.0, 10.0, -8900.0, -10000.0, 9008.0, 0.0,
-        ]);
-
-        let lu_ret = a.lu().unwrap();
-
-        let exp_p =
-            SMatrix::<f64, 3, 3>::from_row_slice(&[0.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 0.0]);
-
-        let exp_l = SMatrix::<f64, 3, 3>::from_row_slice(&[
-            1.00000000e+00,
-            0.00000000e+00,
-            0.00000000e+00,
-            -5.00000000e-01,
-            1.00000000e+00,
-            0.00000000e+00,
-            -1.00000000e-04,
-            4.43265574e-01,
-            1.00000000e+00,
-        ]);
-
-        let exp_u = SMatrix::<f64, 3, 3>::from_row_slice(&[
-            -10000.0,
-            9008.0,
-            0.0,
-            0.0,
-            4514.0,
-            -8900.0,
-            0.0,
-            0.0,
-            6945.06360656,
-        ]);
-
-        for i in 0..9 {
-            assert_relative_eq!(lu_ret.p[i], exp_p[i], max_relative=1.0e-8);
-            assert_relative_eq!(lu_ret.l[i], exp_l[i], max_relative=1.0e-8);
-            assert_relative_eq!(lu_ret.u[i], exp_u[i], max_relative=1.0e-8);
-        }
-    }
-
-    #[test]
-    fn test_diagonal_dominant() 
-    {
-
-        let a = SMatrix::<f64, 4, 4>::from_row_slice(&[
-            100000.0, 10.0,      56.0,      10.0,
-            -69.0,      1.56e6,  3.0,       -9.0,
-            0.0,        0.0,       -5.6e-5, -700.0, 
-            890.0,   0.0, -7899.0, 8.0e5
-        ]);
-
-        let lu_ret = a.lu().unwrap();
-
-        let exp_p = [1., 0., 0., 0.,
-                0., 1., 0., 0.,
-                0., 0., 0., 1.,
-                0., 0., 1., 0.];
-
-        let exp_l = SMatrix::<f64, 4, 4>::from_row_slice(&[
-            1.00000000e+00,  0.00000000e+00 , 0.00000000e+00,  0.00000000e+00,
-            -6.90000000e-04,  1.00000000e+00,  0.00000000e+00,  0.00000000e+00,
-            8.90000000e-03, -5.70512818e-08,  1.00000000e+00,  0.00000000e+00,
-            0.00000000e+00,  0.00000000e+00,  7.08905771e-09,  1.00000000e+00]);
-
-        let exp_u = SMatrix::<f64, 4, 4>::from_row_slice(&[
-                 1.00000000e+05,  1.00000000e+01,  5.60000000e+01,  1.00000000e+01,
-                 0.00000000e+00, 1.56000001e+06,  3.03864000e+00, -8.99310000e+00,
-                 0.00000000e+00,  0.00000000e+00, -7.89949840e+03,  7.99999911e+05,
-                 0.00000000e+00,  0.00000000e+00,  0.00000000e+00, -7.00005671e+02]);
-
-
-        for i in 0..16 {
-            assert_relative_eq!(lu_ret.p[i], exp_p[i], max_relative=1.0e-8);
-            assert_relative_eq!(lu_ret.l[i], exp_l[i], max_relative=1.0e-8);
-            assert_relative_eq!(lu_ret.u[i], exp_u[i], max_relative=1.0e-8);
-        }
-    }
 }
 //}}}
